Harden hot paths: cache Azure AD tokens, guard against cache stampede, widen vote counters to BIGINT
- AzureSqlAuthProvider: share one DefaultAzureCredential and cache its
  token until near expiry instead of re-probing credential sources and
  round-tripping to Entra ID on every Managed Identity connection open
  (every vote), which was dominating per-vote latency under load.
- VoteApiController: guard GET /api/vote/report against a cache
  stampede (concurrent misses now piggy-back on one DB fetch instead of
  each hitting the database), and offload the per-vote CPU/memory/disk
  simulation onto a background thread so it doesn't run on the async
  continuation.
- LoadConfigApiController: reuse a single repo instance for the
  update+refetch instead of opening two separate connections.
- VoteReport's Yes/No/Total/PayloadCount are now long/BIGINT end to
  end (model, all four repositories, and the MSSQL/PostgreSQL
  stored routines) instead of capping at ~2.1 billion, consistent with
  IDVote already being BIGINT for this kind of long-running load test.

+++ ScaleTrigger/Models/Vote.cs
@@ -10,12 +10,12 @@ namespace ScaleTrigger.Models
     /// <summary>Fully computed by the database - no percentage math happens in application code.</summary>
     public class VoteReport
     {
-        public int Yes { get; set; }
-        public int No { get; set; }
-        public int Total { get; set; }
+        public long Yes { get; set; }
+        public long No { get; set; }
+        public long Total { get; set; }
         public decimal YesPercent { get; set; }
         public decimal NoPercent { get; set; }
-        public int PayloadCount { get; set; }
+        public long PayloadCount { get; set; }
         public long PayloadTotalBytes { get; set; }
     }
